Address the comments on config, notifyBPC and upmerge, etc.
